refactored terminal functions and splited into two files
