--- out
@@ -1,10 +1,0 @@
-gnl
-includes
-libft
-Makefile
-minishell
-objs
-out
-README.rd
-rl.supp
-sources
